requirements.txt
anthropic
requests
beautifulsoup4
pyyaml
python-dotenv
schedule
